dto updates & item order api partial working

+++ src/main/java/com/example/demo/repositoriy/OrderRepository.java
@@ -4,6 +4,6 @@ import com.example.demo.model.Order;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface OrderRepository extends CrudRepository<Order, Integer>{
+public interface OrderRepository extends CrudRepository<Order, Long>{
     
 }
